+++ lib/src/mustache.dart
@@ -5,13 +5,10 @@ render(String template, Object context,
     Delimiter delimiter,
     String ident: EMPTY_STRING,
     StringSink out,
-    bool errorOnMissingProperty: false,
-    bool assumeNullNonExistingProperty: true}) {
+    bool errorOnMissingProperty: false}) {
   return compile(template,
           partial: partial, delimiter: delimiter, ident: ident)(context,
-      out: out,
-      errorOnMissingProperty: errorOnMissingProperty,
-      assumeNullNonExistingProperty: assumeNullNonExistingProperty);
+      out: out, errorOnMissingProperty: errorOnMissingProperty);
 }
 
 TemplateRenderer compile(String template,