+++ CHANGELOG.md
@@ -29,7 +29,7 @@ sense with dart. Use a getter instead.
 ## 1.0.9 (2015-02-13)
 
 * throw exception on missing property (helps debugging and tracking down errors) [issue](https://github.com/valotas/mustache4dart/issues/36)
-* introduced assumeNullNonExistingProperty (the difference between a null field and a non-existent field) [issue](https://github.com/valotas/mustache4dart/issues/41)
+* introduced assumeNullNonExistingProperty (the difference between a null field and a non-existent field [issue](https://github.com/valotas/mustache4dart/issues/41)
 * Provide lambdas with the current nested context when they have two parameters (lambdas can now render their contents when inside of a loop)  [issue](https://github.com/valotas/mustache4dart/issues/39)
 
 ## 1.0.8 (2015-02-01)